resource_loop/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.8
django-daraja==1.3.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
phonenumbers==9.0.19
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
python-decouple==3.8
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
celery==5.3.6
redis==5.0.1
djangorestframework==3.14.0
django-anymail[brevo]==10.3
